Separate APIs into separate structs and make Auth scope agnostic
